Move Precision enum declaration to global space
Also, fix spelling/grammar in function name.
