+++ test/unit_test/evaluator/test_olive_evaluator.py
@@ -445,7 +445,7 @@ class TestOliveEvaluatorConfig:
             OliveEvaluatorConfig(metrics=metric_config)
 
     @pytest.mark.parametrize(
-        ("metric_args", "is_accuracy_drop_tolerance"),
+        ("metric_args", "is_accuracy_drop_tolerant"),
         [
             ([{"args": [AccuracySubType.ACCURACY_SCORE], "kwargs": {}}], False),
             ([{"args": [AccuracySubType.ACCURACY_SCORE], "kwargs": {"goal_type": "min-improvement"}}], False),
@@ -454,10 +454,10 @@ class TestOliveEvaluatorConfig:
             ([{"args": [AccuracySubType.ACCURACY_SCORE], "kwargs": {"goal_type": "percent-max-degradation"}}], True),
         ],
     )
-    def test_is_accuracy_drop_tolerance(self, metric_args, is_accuracy_drop_tolerance):
+    def test_is_accuracy_drop_tolerant(self, metric_args, is_accuracy_drop_tolerant):
         evaluator_config = [get_accuracy_metric(*m_arg["args"], **m_arg["kwargs"]) for m_arg in metric_args]
         evaluator_config_instance = OliveEvaluatorConfig(metrics=evaluator_config)
-        assert evaluator_config_instance.is_accuracy_drop_tolerance == is_accuracy_drop_tolerance
+        assert evaluator_config_instance.is_accuracy_drop_tolerant == is_accuracy_drop_tolerant
 
     @patch("olive.common.import_lib.import_user_module")
     @patch("olive.evaluator.registry.Registry.get")
